refactor: JPA 영속화 객체들 jpa 패키지로 이전
