+++ src/main/java/dandi/dandi/comment/adapter/persistence/CommentRepository.java
@@ -9,9 +9,9 @@ public interface CommentRepository extends JpaRepository<CommentJpaEntity, Long>
 
     @Query("SELECT c FROM CommentJpaEntity c WHERE c.postId = :postId "
             + "AND c.memberId NOT IN "
-            + "(SELECT mb.blockedMemberId FROM MemberBlockJpaEntity mb WHERE mb.blockingMemberId = :memberId ) "
+            + "(SELECT mb.blockedMemberId FROM MemberBlockEntity mb WHERE mb.blockingMemberId = :memberId ) "
             + "AND c.memberId NOT IN "
-            + "(SELECT mb.blockingMemberId FROM MemberBlockJpaEntity mb WHERE mb.blockedMemberId = :memberId )"
+            + "(SELECT mb.blockingMemberId FROM MemberBlockEntity mb WHERE mb.blockedMemberId = :memberId )"
             + "AND c.id NOT IN "
             + "(SELECT cr.commentId FROM CommentReportJpaEntity cr WHERE cr.memberId = :memberId)")
     Slice<CommentJpaEntity> findByPostId(Long memberId, Long postId, Pageable pageable);
