+++ src/test/java/dandi/dandi/member/MemberTestFixture.java
@@ -1,6 +1,6 @@
 package dandi.dandi.member;
 
-import dandi.dandi.member.adapter.out.persistence.MemberJpaEntity;
+import dandi.dandi.member.adapter.out.persistence.jpa.MemberEntity;
 import dandi.dandi.member.domain.District;
 import dandi.dandi.member.domain.Location;
 import dandi.dandi.member.domain.Member;
@@ -19,7 +19,7 @@ public class MemberTestFixture {
             MEMBER_ID, OAUTH_ID, NICKNAME, new Location(0.0, 0.0, DISTRICT), INITIAL_PROFILE_IMAGE_URL);
     public static final Member MEMBER2 = new Member(
             2L, "oAuthId2", "nickname2", new Location(0.0, 0.0, DISTRICT), INITIAL_PROFILE_IMAGE_URL);
-    public static final MemberJpaEntity MEMBER_JPA_ENTITY = MemberJpaEntity.fromMember(new Member(
+    public static final MemberEntity MEMBER_JPA_ENTITY = MemberEntity.fromMember(new Member(
             MEMBER_ID, OAUTH_ID, NICKNAME, new Location(0.0, 0.0, DISTRICT), INITIAL_PROFILE_IMAGE_URL));
 
 }
